requirements.txt
customtkinter>=5.2.0
watchdog>=2.3.1
pillow>=9.5.0
pystray>=0.19.4
psutil>=5.9.5
python-dateutil>=2.8.2